+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+- Fixes missing serialization of `Map`s given to `Heaven.prototype.update`.
+
 ## 0.12.0 (Sep 17, 2021)
 - No longer instantiates models from attributes given to `Heaven.prototype.create`.  
   If you do need models for client-side default attributes, for example, instantiate them before calling `Heaven.prototype.create`:
@@ -24,6 +27,7 @@
   ```javascript
   require("heaven") == require("heaven/async")
   ```
+
 ## 0.11.3 (Sep 17, 2021)
 Republished v0.11.1 as v0.11.3 as unpublishing v0.11.2 from NPM was no longer possible.